+++ README.md
@@ -1,11 +1,10 @@
 # project-react-6-shongs27
 
 - [v][pr3] db 데이터에 따라 블로그 글 목록 렌더링
-- [][pr4] post 데이터를 back서버가 아니라 정적컨텐츠로 작성
-
-- [ ][pr4] 글 목록에서 글을 누르면 그 페이지 항목 렌더링
-  - 각 페이지 테스트 완성
+- [ ][pr4] 글 목록에서 해당항목을 누르면 해당 페이지 내용 렌더링
   - 더보기 누르면 상세 페이지 만들기
+    중첩라우팅으로 구현
+    https://www.daleseo.com/react-router-nested/
 
 * 최근 포스트 / 인기포스트 항목 완성?
