+++ src/PagesContainer.jsx
@@ -5,6 +5,7 @@ import ReactPage from './pages/ReactPage';
 // import ComputerPage from './pages/ComputerPage';
 // import AlgorithmPage from './pages/AlgorithmPage';
 // import SomethingPage from './pages/SomethingPage';
+import PageDetail from './components/PageDetail';
 
 import TopContainer from './commons/top/TopContainer';
 import Footer from './commons/footer/FooterContainer';
@@ -30,13 +31,15 @@ export default function PagesContainer() {
         <MiddleContainer>
           <Routes>
             <Route path="/" element={<MainPage />} />
+            <Route path="/:id" element={<PageDetail />} />
+
             <Route path="/js" element={<JsPage />} />
             <Route path="/react" element={<ReactPage />} />
             {/* <Route path="/computer" element={<ComputerPage />} />
             <Route path="/algorithm" element={<AlgorithmPage />} />
             <Route path="/something" element={<SomethingPage />} /> */}
 
-            <Route path="*" element={<MainPage />} />
+            <Route path="*" element={<div>고장났어요</div>} />
           </Routes>
         </MiddleContainer>
       </Container>
