+++ src/components/PageContent.jsx
@@ -1,0 +1,46 @@
+import { Link, Routes, Route, useParams } from 'react-router-dom';
+
+import styled from '@emotion/styled';
+
+const Article = styled.article({
+  minHeight: '50vh',
+  marginBottom: '3em',
+
+  '& h3': {
+    padding: '1em 0',
+    borderBottom: '1px solid rgb(230, 230, 230)',
+  },
+});
+
+const Item = styled.div({
+  borderBottom: '1px solid rgb(230, 230, 230)',
+  paddingTop: '30px',
+  paddingBottom: '22px',
+});
+
+export default function PageDetail({ articleTitle, pageContents }) {
+  return (
+    <Article id="content">
+      <h3>{articleTitle}</h3>
+      <section className="inner">
+        {pageContents &&
+          pageContents.map(({ id, title, content }) => {
+            //excerpt 기능
+            if (content.length > 200) {
+              content = content.substring(0, 200) + '...';
+            }
+
+            return (
+              <Item key={id} className="post_item">
+                <span>{title}</span>
+                <p className="post_content" style={{ color: '#666' }}>
+                  {content}
+                </p>
+                <Link to="/3">더보기 &gt;</Link>
+              </Item>
+            );
+          })}
+      </section>
+    </Article>
+  );
+}
